Migrate all "renderer" files into proper directory
