--- src/component/Connection/types.ts
@@ -1,7 +1,0 @@
-export type ConnectionObject = {
-  name: string;
-  host: string;
-  port: number;
-  user: string;
-  password: string;
-};
