+++ src/configuration/index.ts
@@ -2,8 +2,8 @@ import { dialog, safeStorage } from 'electron';
 import { resolve } from 'node:path';
 import { existsSync, mkdirSync, readFileSync, writeFile } from 'node:fs';
 import envPaths from 'env-paths';
-import { ConnectionObject } from '../component/Connection/types';
-import { DEFAULT_THEME } from '../theme';
+import { ConnectionObject } from '../sql/types';
+import { DEFAULT_THEME } from './themes';
 import {
   Configuration,
   EncryptedConnectionObject,
