+++ src/renderer.ts
@@ -26,7 +26,7 @@
  * ```
  */
 
-import './index.css';
+import './renderer/index.css';
 
 // Add this to the end of the existing file
-import './app';
+import './renderer/app';
