+++ src/renderer/component/Cell.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import Cell from './Cell';
-import { DatabaseContext } from '../contexts/DatabaseContext';
-import { ConnectionContext } from '../contexts/ConnectionContext';
+import { DatabaseContext } from '../../contexts/DatabaseContext';
+import { ConnectionContext } from '../../contexts/ConnectionContext';
 import { Types } from 'mysql';
 
 const meta: Meta<typeof Cell> = {
