+++ src/renderer/component/Connection/types.ts
