+++ src/renderer/component/Style/ModalLike.tsx
