+++ src/renderer/routes/connect/edit.$connectionName.tsx
@@ -2,7 +2,7 @@ import { useParams } from 'react-router';
 import invariant from 'tiny-invariant';
 import ModalLike from '../../component/Style/ModalLike';
 import ConnectionForm from '../../component/Connection/ConnectionForm';
-import { useConfiguration } from '../../contexts/ConfigurationContext';
+import { useConfiguration } from '../../../contexts/ConfigurationContext';
 
 export default function Edit() {
   const { configuration } = useConfiguration();
