+++ src/renderer/routes/connections.$connectionName.$databaseName.$tableName.tsx
