+++ src/sql/types.ts
@@ -15,3 +15,15 @@ export type QueryReturnType =
 export type QueryResult<T extends QueryReturnType = QueryReturnType> = Promise<
   [T, FieldPacket[]]
 >;
+
+/**
+ * Represet a connection object that is stored in the configuration file.
+ * It is also used to connect to the database.
+ */
+export type ConnectionObject = {
+  name: string;
+  host: string;
+  port: number;
+  user: string;
+  password: string;
+};
